requirements.txt
streamlit==0.89.0
streamlit-folium==0.4.0
folium==0.12.1
pandas==1.1.5
geopandas==0.6.1
numpy==1.19.5
pydeck==0.7.0
pyngrok==5.1.0
altair==4.1.0
Pillow==7.1.2
requests==2.23.0
requests-oauthlib==1.3.0
urllib3==1.24.3